+++ test/haskell-language-server/stack.nix
@@ -1,16 +1,16 @@
-{ testSrc, haskell-nix, evalPackages, buildPackages, compiler-nix-name }:
+{ lib, testSrc, haskell-nix, evalPackages, buildPackages, compiler-nix-name }:
 
 (buildPackages.haskell-nix.project {
   inherit compiler-nix-name;
   src = evalPackages.fetchgit {
     url = "https://github.com/haskell/haskell-language-server.git";
     fetchSubmodules = true;
-    rev = "d2654185eef1b0d703cebc694e85438e20600e37";
-    sha256 = "0s0k2i0imkcn9zykhrlqq0r4ssv25mwbpdyc675xkzgl1vj1l8kd";
+    rev = "8d37a64f359950f3b340ec50ed419742758ebf0c";
+    sha256 = "0w2v133yk0bhnbsj7hnsh5w2kzyd86r6xrkx4q167jyrfq8sa1qg";
   };
-  projectFileName = "stack-${haskell-nix.compiler.${compiler-nix-name}.version}.yaml";
+  projectFileName = "stack-${haskell-nix.compiler.${if compiler-nix-name == "ghc884" then "ghc883" else compiler-nix-name}.version}.yaml";
+  modules = [{ config.compiler.nix-name = lib.mkForce compiler-nix-name; }];
   sha256map = {
-    "https://github.com/wz1000/shake.git"."fb3859dca2e54d1bbb2c873e68ed225fa179fbef" = "0sa0jiwgyvjsmjwpfcpvzg2p7277aa0dgra1mm6afh2rfnjphz8z";
-    "https://github.com/peti/cabal-plan.git"."894b76c0b6bf8f7d2f881431df1f13959a8fce87" = "06iklj51d9kh9bhc42lrayypcpgkjrjvna59w920ln41rskhjr4y";
+    "https://github.com/DanielG/cabal-helper.git"."79a5608778493bf32e74b54bbf1ea2729941e50f" = "1jsiwg94yy8pwhzi3z6ayja9qdgf7fl6xn1h9z681j6lhbx225f8";
   };
 }).haskell-language-server.components.exes.haskell-language-server